Moved ZT models to own folder and added documentation
