+++ monkey/monkey_island/cc/models/zero_trust/test_event.py
@@ -1,7 +1,7 @@
 from mongoengine import ValidationError
 
 from common.data.zero_trust_consts import EVENT_TYPE_ISLAND
-from monkey_island.cc.models.event import Event
+from monkey_island.cc.models.zero_trust.event import Event
 from monkey_island.cc.testing.IslandTestCase import IslandTestCase
